+++ js/tutorials/photo-albums-in-react/.prettierrc
@@ -1,3 +1,3 @@
 {
-  "singleQuote": true
+	"singleQuote": true
 }
