+++ js/tutorials/photo-albums-in-react/_steps/src/components/AlbumDetails.js/p4s5.js.jekyll
@@ -1,22 +1,7 @@
 ```jsx
-{% raw %}import { API, Storage } from 'aws-amplify';
+{% raw %}import { API } from 'aws-amplify';
 import React, { useEffect, useState } from 'react';
-import {
-  Button,
-  Dimmer,
-  Header,
-  Loader,
-  Message,
-  Segment,
-  Form
-} from 'semantic-ui-react';
-
-import {
-  onCreatePhoto,
-  onDeletePhoto,
-  onUpdateAlbum,
-  onUpdatePhoto
-} from '../graphql/subscriptions';
+import { Dimmer, Header, Loader, Message, Segment } from 'semantic-ui-react';
 
 import DeleteAlbum from './DeleteAlbum';
 import PhotoList from './PhotoList';
@@ -24,83 +9,13 @@ import { getAlbum } from '../graphql/queries';
 import useAuth from '../useAuth';
 
 function AlbumDetails({ album }) {
-  const { owner } = useAuth();
-  const [photos, setPhotos] = useState(album.photos.items);
-
-  useEffect(() => {
-    const subscription = API.graphql({
-      authMode: owner ? 'AMAZON_COGNITO_USER_POOLS' : 'AWS_IAM',
-      query: onCreatePhoto,
-      variables: { owner }
-    }).subscribe({
-      next(payload) {
-        const photo = payload.value.data.onCreatePhoto;
-
-        setPhotos(prevPhotos => [photo, ...prevPhotos]);
-      }
-    });
-
-    return () => subscription.unsubscribe();
-  }, [owner]);
-
-  useEffect(() => {
-    const subscription = API.graphql({
-      authMode: owner ? 'AMAZON_COGNITO_USER_POOLS' : 'AWS_IAM',
-      query: onDeletePhoto,
-      variables: { owner }
-    }).subscribe({
-      next(payload) {
-        const photo = payload.value.data.onDeletePhoto;
-
-        Storage.remove(photo.fullsize.key);
-
-        setPhotos(prevPhotos => {
-          return prevPhotos.filter(prevPhoto => prevPhoto.id !== photo.id);
-        });
-      }
-    });
-
-    return () => subscription.unsubscribe();
-  }, [owner]);
-
-  useEffect(() => {
-    const subscription = API.graphql({
-      authMode: owner ? 'AMAZON_COGNITO_USER_POOLS' : 'AWS_IAM',
-      query: onUpdatePhoto,
-      variables: { owner }
-    }).subscribe({
-      next(payload) {
-        const photo = payload.value.data.onUpdatePhoto;
-
-        setPhotos(prevPhotos => {
-          return prevPhotos.map(prevPhoto => {
-            return prevPhoto.id === photo.id ? photo : prevPhoto;
-          });
-        });
-      }
-    });
-
-    return () => subscription.unsubscribe();
-  }, [owner]);
+  const photos = album.photos.items;
 
   return (
     <Segment.Group>
       <Segment clearing>
         <Header data-test="album-header" floated="left" size="huge">
-          <Header.Content>album.name</Header.Content>
-
-          {owner && (
-            <Header.Subheader>
-              <Button
-                basic
-                circular
-                content="Edit"
-                icon="pencil"
-                size="mini"
-                style={{ position: 'relative', top: '0.5em' }}
-              />
-            </Header.Subheader>
-          )}
+          <Header.Content>{album.name}</Header.Content>
         </Header>
       </Segment>
 
@@ -135,20 +50,6 @@ export default function AlbumDetailsLoader(props) {
     });
   }, [id, owner]);
 
-  useEffect(() => {
-    const subscription = API.graphql({
-      authMode: owner ? 'AMAZON_COGNITO_USER_POOLS' : 'AWS_IAM',
-      query: onUpdateAlbum,
-      variables: { owner }
-    }).subscribe({
-      next(payload) {
-        setAlbum(payload.value.data.onUpdateAlbum);
-      }
-    });
-
-    return () => subscription.unsubscribe();
-  }, [id, owner]);
-
   if (isLoading) {
     return (
       <Dimmer active inverted page>
@@ -165,7 +66,7 @@ export default function AlbumDetailsLoader(props) {
     );
   }
 
-  return <AlbumDetails album={album} key={`${album.id}-${album.version}`} />;
+  return <AlbumDetails album={album} />;
 }
 {% endraw %}
 ```
