+++ js/tutorials/photo-albums-in-react/_steps/src/components/AlbumDetails.js/p4s6.js.patch.jekyll
@@ -1,7 +1,22 @@
 ```diff
 {% raw %}--- a/src/components/AlbumDetails.js
 +++ b/src/components/AlbumDetails.js
-@@ -20,10 +20,13 @@ import {
+@@ -1,20 +1,88 @@
+ import { API } from 'aws-amplify';
+ import React, { useEffect, useState } from 'react';
+-import { Dimmer, Header, Loader, Message, Segment } from 'semantic-ui-react';
++import {
++  Button,
++  Dimmer,
++  Header,
++  Loader,
++  Message,
++  Segment,
++  Form
++} from 'semantic-ui-react';
++
++import { onUpdateAlbum } from '../graphql/subscriptions';
+ 
  import DeleteAlbum from './DeleteAlbum';
  import PhotoList from './PhotoList';
  import { getAlbum } from '../graphql/queries';
@@ -9,15 +24,10 @@
  import useAuth from '../useAuth';
  
  function AlbumDetails({ album }) {
-   const { owner } = useAuth();
++  const { owner } = useAuth();
 +  const [isEditing, setIsEditing] = useState(false);
 +  const [isUpdating, setIsUpdating] = useState(false);
-   const [photos, setPhotos] = useState(album.photos.items);
- 
-   useEffect(() => {
-@@ -82,19 +85,60 @@ function AlbumDetails({ album }) {
-     return () => subscription.unsubscribe();
-   }, [owner]);
+   const photos = album.photos.items;
  
 +  const handleCancel = () => {
 +    setIsEditing(!isEditing);
@@ -45,7 +55,7 @@
      <Segment.Group>
        <Segment clearing>
          <Header data-test="album-header" floated="left" size="huge">
--          <Header.Content>album.name</Header.Content>
+-          <Header.Content>{album.name}</Header.Content>
 +          <Header.Content>
 +            {isEditing ? (
 +              <Form onSubmit={handleSave}>
@@ -65,18 +75,49 @@
 +              album.name
 +            )}
 +          </Header.Content>
- 
-           {owner && (
-             <Header.Subheader>
-               <Button
-                 basic
-                 circular
--                content="Edit"
--                icon="pencil"
++
++          {owner && (
++            <Header.Subheader>
++              <Button
++                basic
++                circular
 +                content={isEditing ? 'Cancel' : 'Edit'}
 +                icon={isEditing ? 'close' : 'pencil'}
 +                onClick={handleCancel}
-                 size="mini"
-                 style={{ position: 'relative', top: '0.5em' }}
-               />{% endraw %}
++                size="mini"
++                style={{ position: 'relative', top: '0.5em' }}
++              />
++            </Header.Subheader>
++          )}
+         </Header>
+       </Segment>
+ 
+@@ -49,6 +117,20 @@ export default function AlbumDetailsLoader(props) {
+     });
+   }, [id, owner]);
+ 
++  useEffect(() => {
++    const subscription = API.graphql({
++      authMode: owner ? 'AMAZON_COGNITO_USER_POOLS' : 'AWS_IAM',
++      query: onUpdateAlbum,
++      variables: { owner }
++    }).subscribe({
++      next(payload) {
++        setAlbum(payload.value.data.onUpdateAlbum);
++      }
++    });
++
++    return () => subscription.unsubscribe();
++  }, [id, owner]);
++
+   if (isLoading) {
+     return (
+       <Dimmer active inverted page>
+@@ -65,5 +147,5 @@ export default function AlbumDetailsLoader(props) {
+     );
+   }
+ 
+-  return <AlbumDetails album={album} />;
++  return <AlbumDetails album={album} key={`${album.id}-${album.version}`} />;
+ }{% endraw %}
 ```
