+++ js/tutorials/photo-albums-in-react/_steps/src/components/AlbumDetails.js/p6s1.js
@@ -1,4 +1,4 @@
-import { API, Storage } from 'aws-amplify';
+import { API } from 'aws-amplify';
 import React, { useEffect, useState } from 'react';
 import {
   Button,
@@ -10,12 +10,7 @@ import {
   Form
 } from 'semantic-ui-react';
 
-import {
-  onCreatePhoto,
-  onDeletePhoto,
-  onUpdateAlbum,
-  onUpdatePhoto
-} from '../graphql/subscriptions';
+import { onUpdateAlbum } from '../graphql/subscriptions';
 
 import DeleteAlbum from './DeleteAlbum';
 import PhotoList from './PhotoList';
@@ -28,63 +23,7 @@ function AlbumDetails({ album }) {
   const { owner } = useAuth();
   const [isEditing, setIsEditing] = useState(false);
   const [isUpdating, setIsUpdating] = useState(false);
-  const [photos, setPhotos] = useState(album.photos.items);
-
-  useEffect(() => {
-    const subscription = API.graphql({
-      authMode: owner ? 'AMAZON_COGNITO_USER_POOLS' : 'AWS_IAM',
-      query: onCreatePhoto,
-      variables: { owner }
-    }).subscribe({
-      next(payload) {
-        const photo = payload.value.data.onCreatePhoto;
-
-        setPhotos(prevPhotos => [photo, ...prevPhotos]);
-      }
-    });
-
-    return () => subscription.unsubscribe();
-  }, [owner]);
-
-  useEffect(() => {
-    const subscription = API.graphql({
-      authMode: owner ? 'AMAZON_COGNITO_USER_POOLS' : 'AWS_IAM',
-      query: onDeletePhoto,
-      variables: { owner }
-    }).subscribe({
-      next(payload) {
-        const photo = payload.value.data.onDeletePhoto;
-
-        Storage.remove(photo.fullsize.key);
-
-        setPhotos(prevPhotos => {
-          return prevPhotos.filter(prevPhoto => prevPhoto.id !== photo.id);
-        });
-      }
-    });
-
-    return () => subscription.unsubscribe();
-  }, [owner]);
-
-  useEffect(() => {
-    const subscription = API.graphql({
-      authMode: owner ? 'AMAZON_COGNITO_USER_POOLS' : 'AWS_IAM',
-      query: onUpdatePhoto,
-      variables: { owner }
-    }).subscribe({
-      next(payload) {
-        const photo = payload.value.data.onUpdatePhoto;
-
-        setPhotos(prevPhotos => {
-          return prevPhotos.map(prevPhoto => {
-            return prevPhoto.id === photo.id ? photo : prevPhoto;
-          });
-        });
-      }
-    });
-
-    return () => subscription.unsubscribe();
-  }, [owner]);
+  const photos = album.photos.items;
 
   const handleCancel = () => {
     setIsEditing(!isEditing);
