+++ js/tutorials/photo-albums-in-react/_steps/src/components/AlbumDetails.js/p6s4.js.jekyll
@@ -1,5 +1,5 @@
 ```jsx
-{% raw %}import { API } from 'aws-amplify';
+{% raw %}import { API, Storage } from 'aws-amplify';
 import React, { useEffect, useState } from 'react';
 import {
   Button,
@@ -11,10 +11,16 @@ import {
   Form
 } from 'semantic-ui-react';
 
-import { onUpdateAlbum } from '../graphql/subscriptions';
+import {
+  onCreatePhoto,
+  onDeletePhoto,
+  onUpdateAlbum,
+  onUpdatePhoto
+} from '../graphql/subscriptions';
 
 import DeleteAlbum from './DeleteAlbum';
 import PhotoList from './PhotoList';
+import PhotosUploader from './PhotosUploader';
 import { getAlbum } from '../graphql/queries';
 import { updateAlbum } from '../graphql/mutations';
 import useAuth from '../useAuth';
@@ -23,7 +29,63 @@ function AlbumDetails({ album }) {
   const { owner } = useAuth();
   const [isEditing, setIsEditing] = useState(false);
   const [isUpdating, setIsUpdating] = useState(false);
-  const photos = album.photos.items;
+  const [photos, setPhotos] = useState(album.photos.items);
+
+  useEffect(() => {
+    const subscription = API.graphql({
+      authMode: owner ? 'AMAZON_COGNITO_USER_POOLS' : 'AWS_IAM',
+      query: onCreatePhoto,
+      variables: { owner }
+    }).subscribe({
+      next(payload) {
+        const photo = payload.value.data.onCreatePhoto;
+
+        setPhotos(prevPhotos => [photo, ...prevPhotos]);
+      }
+    });
+
+    return () => subscription.unsubscribe();
+  }, [owner]);
+
+  useEffect(() => {
+    const subscription = API.graphql({
+      authMode: owner ? 'AMAZON_COGNITO_USER_POOLS' : 'AWS_IAM',
+      query: onDeletePhoto,
+      variables: { owner }
+    }).subscribe({
+      next(payload) {
+        const photo = payload.value.data.onDeletePhoto;
+
+        Storage.remove(photo.fullsize.key);
+
+        setPhotos(prevPhotos => {
+          return prevPhotos.filter(prevPhoto => prevPhoto.id !== photo.id);
+        });
+      }
+    });
+
+    return () => subscription.unsubscribe();
+  }, [owner]);
+
+  useEffect(() => {
+    const subscription = API.graphql({
+      authMode: owner ? 'AMAZON_COGNITO_USER_POOLS' : 'AWS_IAM',
+      query: onUpdatePhoto,
+      variables: { owner }
+    }).subscribe({
+      next(payload) {
+        const photo = payload.value.data.onUpdatePhoto;
+
+        setPhotos(prevPhotos => {
+          return prevPhotos.map(prevPhoto => {
+            return prevPhoto.id === photo.id ? photo : prevPhoto;
+          });
+        });
+      }
+    });
+
+    return () => subscription.unsubscribe();
+  }, [owner]);
 
   const handleCancel = () => {
     setIsEditing(!isEditing);
@@ -85,6 +147,7 @@ function AlbumDetails({ album }) {
             </Header.Subheader>
           )}
         </Header>
+        <PhotosUploader album={album} />
       </Segment>
 
       <Segment data-test="album-content" tertiary>
