+++ js/tutorials/photo-albums-in-react/_steps/src/components/PhotoDetails.js/p6s2.js
@@ -1,23 +1,25 @@
 import React from 'react';
+import { Link, useParams } from 'react-router-dom';
 import { Card, Label } from 'semantic-ui-react';
 
 import Photo from './Photo';
 
 export default function PhotoDetails({ photo }) {
+  const { albumId } = useParams();
   const [description, labels] = photo;
 
   return (
     <Card data-test="photo-card" raised>
-      <Photo photo={photo} thumbnail />
+      <Link
+        data-test="photo-card-thumbnail"
+        style={{ background: 'black' }}
+        to={`/albums/${albumId}/photos/${photo.id}`}
+      >
+        <Photo photo={photo} thumbnail />
+      </Link>
 
       <Card.Content data-test="photo-card-description">
-        <Card.Meta
-          style={{
-            whiteSpace: 'pre-line'
-          }}
-        >
-          {description}
-        </Card.Meta>
+        <Card.Meta style={{ whiteSpace: 'pre-line' }}>{description}</Card.Meta>
       </Card.Content>
 
       {labels.length ? (
