+++ js/tutorials/photo-albums-in-react/_steps/src/components/PhotoDetails.js/p6s2.js.patch.jekyll
@@ -1,20 +1,28 @@
 ```diff
 {% raw %}--- a/src/components/PhotoDetails.js
 +++ b/src/components/PhotoDetails.js
-@@ -1,11 +1,15 @@
+@@ -1,11 +1,23 @@
  import React from 'react';
++import { Link, useParams } from 'react-router-dom';
  import { Card, Label } from 'semantic-ui-react';
  
 +import Photo from './Photo';
 +
  export default function PhotoDetails({ photo }) {
++  const { albumId } = useParams();
    const [description, labels] = photo;
  
    return (
      <Card data-test="photo-card" raised>
-+      <Photo photo={photo} thumbnail />
++      <Link
++        data-test="photo-card-thumbnail"
++        style={{ background: 'black' }}
++        to={`/albums/${albumId}/photos/${photo.id}`}
++      >
++        <Photo photo={photo} thumbnail />
++      </Link>
 +
        <Card.Content data-test="photo-card-description">
-         <Card.Meta
-           style={{{% endraw %}
+         <Card.Meta style={{ whiteSpace: 'pre-line' }}>{description}</Card.Meta>
+       </Card.Content>{% endraw %}
 ```
