+++ js/tutorials/photo-albums-in-react/_steps/src/components/PhotoDetails.js/p6s3.js.jekyll
@@ -3,8 +3,8 @@
 import React, { useState } from 'react';
 import { Link, useParams } from 'react-router-dom';
 import { Button, Card, Label, Form, TextArea } from 'semantic-ui-react';
-import Photo from './Photo';
 
+import Photo from './Photo';
 import { updatePhoto } from '../graphql/mutations';
 import useAuth from '../useAuth';
 
@@ -68,11 +68,7 @@ export default function PhotoDetails({ photo }) {
             />
           </Form>
         ) : (
-          <Card.Meta
-            style={{
-              whiteSpace: 'pre-line'
-            }}
-          >
+          <Card.Meta style={{ whiteSpace: 'pre-line' }}>
             {description}
           </Card.Meta>
         )}
