+++ js/tutorials/photo-albums-in-react/_steps/src/components/PhotoDetails.js/p6s4.js.patch.jekyll
@@ -1,29 +1,43 @@
 ```diff
 {% raw %}--- a/src/components/PhotoDetails.js
 +++ b/src/components/PhotoDetails.js
-@@ -1,10 +1,11 @@
+@@ -1,10 +1,10 @@
  import { API } from 'aws-amplify';
  import React, { useState } from 'react';
  import { Link, useParams } from 'react-router-dom';
 -import { Button, Card, Label, Form, TextArea } from 'semantic-ui-react';
-+import { Button, Card, Icon, Label, Form, TextArea } from 'semantic-ui-react';
-+
- import Photo from './Photo';
++import { Button, Card, Form, TextArea } from 'semantic-ui-react';
  
+ import Photo from './Photo';
 -import { updatePhoto } from '../graphql/mutations';
 +import { deletePhoto, updatePhoto } from '../graphql/mutations';
  import useAuth from '../useAuth';
  
  export default function PhotoDetails({ photo }) {
-@@ -14,6 +15,7 @@ export default function PhotoDetails({ photo }) {
+@@ -13,11 +13,10 @@ export default function PhotoDetails({ photo }) {
+   const [isEditing, setIsEditing] = useState(false);
    const [isUpdating, setIsUpdating] = useState(false);
    const [description, setDescription] = useState(photo.description);
-   const [labels, setLabels] = useState(photo.labels);
+-  const [labels, setLabels] = useState(photo.labels);
 +  const [isDeleting, setIsDeleting] = useState(false);
  
    const reset = () => {
      setDescription(photo.description);
-@@ -45,6 +47,21 @@ export default function PhotoDetails({ photo }) {
+-    setLabels(photo.labels);
+     setIsEditing(false);
+   };
+ 
+@@ -32,8 +31,7 @@ export default function PhotoDetails({ photo }) {
+           input: {
+             expectedVersion: photo.version,
+             id: photo.id,
+-            description,
+-            labels
++            description
+           }
+         }
+       });
+@@ -45,6 +43,21 @@ export default function PhotoDetails({ photo }) {
      setIsEditing(false);
    };
  
@@ -45,7 +59,28 @@
    return (
      <Card data-test="photo-card" raised>
        <Link
-@@ -127,6 +144,21 @@ export default function PhotoDetails({ photo }) {
+@@ -73,20 +86,6 @@ export default function PhotoDetails({ photo }) {
+         )}
+       </Card.Content>
+ 
+-      {labels.length ? (
+-        <Card.Content extra>
+-          <Card.Description>
+-            <Label.Group>
+-              {labels.map(label => (
+-                <Label key={label} size="tiny">
+-                  {label}
+-                </Label>
+-              ))}
+-            </Label.Group>
+-          </Card.Description>
+-        </Card.Content>
+-      ) : null}
+-
+       {owner && (
+         <Card.Content extra>
+           {!isEditing && (
+@@ -123,6 +122,21 @@ export default function PhotoDetails({ photo }) {
                onClick={reset}
              />
            )}
