+++ js/tutorials/photo-albums-in-react/_steps/src/components/PhotoDetails.js/p7s5.js.jekyll
@@ -5,7 +5,6 @@ import { Link, useParams } from 'react-router-dom';
 import { Button, Card, Icon, Label, Form, TextArea } from 'semantic-ui-react';
 
 import Photo from './Photo';
-
 import { deletePhoto, updatePhoto } from '../graphql/mutations';
 import useAuth from '../useAuth';
 
@@ -94,11 +93,7 @@ export default function PhotoDetails({ photo }) {
             />
           </Form>
         ) : (
-          <Card.Meta
-            style={{
-              whiteSpace: 'pre-line'
-            }}
-          >
+          <Card.Meta style={{ whiteSpace: 'pre-line' }}>
             {description}
           </Card.Meta>
         )}
